# opt-level and max-methods = 1