--- a/src/Latexify.jl
+++ b/src/Latexify.jl
@@ -1,4 +1,12 @@
 module Latexify
+
+if isdefined(Base, :Experimental) && isdefined(Base.Experimental, Symbol("@optlevel"))
+    @eval Base.Experimental.@optlevel 1
+end
+if isdefined(Base, :Experimental) && isdefined(Base.Experimental, Symbol("@max_methods"))
+    @eval Base.Experimental.@max_methods 1
+end
+
 using Requires
 using LaTeXStrings
 using InteractiveUtils