Trying to reduce latency.
After:
```julia
Test Summary: | Pass  Total  Time
macro test    |   25     25  2.6s
Test Summary: | Pass  Total  Time
recipe test   |   13     13  4.4s
Test Summary:  | Pass  Total  Time
latexify tests |   10     10  1.3s
Test Summary:  | Pass  Broken  Total  Time
latexraw tests |  133       1    134  4.8s
Test Summary:    | Pass  Total  Time
latexalign tests |    4      4  1.3s
Test Summary:    | Pass  Total  Time
latexarray tests |   15     15  1.5s
Test Summary:       | Pass  Total  Time
latexequation tests |    2      2  0.1s
Test Summary:      | Pass  Total  Time
latexbracket tests |    2      2  0.0s
Test Summary:     | Pass  Total  Time
latexinline tests |    1      1  0.3s
Test Summary:      | Pass  Total  Time
latextabular tests |   11     11  7.3s
WARNING: using DataFrames.head in module Main conflicts with an existing identifier.
Test Summary: | Pass  Total  Time
mdtable tests |   16     16  2.7s
Test Summary:    | Pass  Broken  Total  Time
DataFrame Plugin |    3       1      4  0.6s
Test Summary:    | Pass  Total  Time
SymEngine Plugin |    1      1  0.5s
```
Before:
```julia
Test Summary: | Pass  Total  Time
macro test    |   25     25  3.5s
Test Summary: | Pass  Total  Time
recipe test   |   13     13  5.4s
Test Summary:  | Pass  Total  Time
latexify tests |   10     10  1.6s
Test Summary:  | Pass  Broken  Total  Time
latexraw tests |  133       1    134  5.9s
Test Summary:    | Pass  Total  Time
latexalign tests |    4      4  2.0s
Test Summary:    | Pass  Total  Time
latexarray tests |   15     15  2.1s
Test Summary:       | Pass  Total  Time
latexequation tests |    2      2  0.1s
Test Summary:      | Pass  Total  Time
latexbracket tests |    2      2  0.1s
Test Summary:     | Pass  Total  Time
latexinline tests |    1      1  0.3s
Test Summary:      | Pass  Total  Time
latextabular tests |   11     11  7.8s
WARNING: using DataFrames.head in module Main conflicts with an existing identifier.
Test Summary: | Pass  Total  Time
mdtable tests |   16     16  3.1s
Test Summary:    | Pass  Broken  Total  Time
DataFrame Plugin |    3       1      4  0.6s
Test Summary:    | Pass  Total  Time
SymEngine Plugin |    1      1  0.5s
```
I was hoping for more.