@chriselrod 2023-01-30
Some performance benchmarks on a big file of latexify tests:
[myb/concrete_only](https://github.com/YingboMa/julia/tree/myb/concrete_only), latexify 0.15.13:
```julia
julia> using Latexify, Pumas, Test, StaticArrays
WARNING: using StaticArrays.pop in module Main conflicts with an existing identifier.

julia> @time include("latexify_recipes.jl")
326.492189 seconds (335.92 M allocations: 15.622 GiB, 42.37% gc time, 57.01% compilation time: <1% of which was recompilation)
Test Passed
```
Julia master, latexify 0.15.13:
```julia
julia> @time include("/home/chriselrod/.julia/dev/Pumas5/test/features/latexify_recipes.jl")
1104.623542 seconds (1.15 G allocations: 78.342 GiB, 29.51% gc time, 88.57% compilation time: 2% of which was recompilation)
Test Passed
```

Note that this Julia branch has no negative performance impact on fast, type stable, code.
It needs more testing, but ODE solves, `@turbo` loops, etc are just as fast on Julia master, while slow code that spends an eternity compiling (like Latexify.jl) is far faster.

I was hoping the tweaks in this PR would also give a 2 or 3x improvement, but alas, we'll have to wait for some form of this branch to make it into Julia, or improve this library.