@chriselrod 2023-01-30
`myb/concrete_only` with this PR:
```julia
julia> @time include("latexify_recipes.jl")
327.410396 seconds (337.46 M allocations: 15.694 GiB, 42.77% gc time, 56.58% compilation time: <1% of which was recompilation)
```
Master with this PR:
```julia
julia> @time include("latexify_recipes.jl")
WARNING: using StaticArrays.pop in module Main conflicts with an existing identifier.
1519.625729 seconds (1.20 G allocations: 82.093 GiB, 24.20% gc time, 91.87% compilation time: 1% of which was recompilation)
```
So, this PR seems to have regressed a lot.