@chriselrod 2023-01-30
Latest Latexify release:
Julia myb/concrete_only:
```julia
julia> @time include("latexify_recipes.jl")
330.552732 seconds (337.50 M allocations: 15.696 GiB, 42.14% gc time, 58.56% compilation time: <1% of which was recompilation)
Test Passed
```

Julia master:
```julia
julia> @time include("latexify_recipes.jl")
1124.050085 seconds (1.15 G allocations: 78.384 GiB, 31.80% gc time, 87.93% compilation time: 2% of which was recompilation)
Test Passed
```