@gustaphe 2023-01-30
Is there somewhere we can read about this? Is the myb branch "official" in that these commands are guaranteed to not mean something else in an upcoming Julia release? Or is there maybe a pull request this should wait for?

`latexify` should not be performance critical, but lower latency is always nice :)