@ChrisRackauckas 2023-01-30
These commands have been around for years. It's what helped Plots.jl improve its load time in like 2020. opt-level=1 is just the same as forcing -O1 on the module, and max-methods=1 is just disabling union splitting optimizations. Just search and you'll find tons about this because they are not new options, just Yingbo's branch improves a few things with them.