@korsbo 2023-01-30
@gustaphe runtime is usually not a problem but there are some real latency issues. The worst ones I've seen is about 30 seconds where there are recipes calling other recipes. That absolutely precludes latexify from being used in the `show` methods in those cases so there are cases where latency is a real problem.