@gustaphe 2023-01-30
Okay, as far as I knew this was related to the `myb` branch. A cursory search mostly returns results about "`max` vs `maximum`" and `Optim.jl` respectively. 

If it's that stable and gives these speedups, LGTM.