@ChrisRackauckas 2023-01-30
This package is missing snoopprecompile: that seems to be a major oversight given its mostly a compile time burdened library?