@korsbo 2023-01-30
I have tried these macros myself, probably back in julia 1.6, but with no tangible effect so I never pushed them. I was hoping that my rewrite (which has been stalled for like 2 years) would fix the issue but I'm not entirely sure that it does either. That rewrite would incur a massive amount of downstream work since it's breaking the current recipe system, hence my tardiness of getting it done.