@korsbo 2023-01-30
Yeah, SnoopPreCompile would certainly be worth a shot. I suspect that much of the bad performance is coming from the recipes which I don't think would be fixed by this. But it could still make an impact and it's worth adding.