@chriselrod 2023-01-30
No, I was noting that the myb/concrete_only Julia branch tweaked inference heuristics, leading to a 3x+ improvement for Latexify.

However, tweaking the options we have available in an official release didn't really seem to help much at all.