@chriselrod 2023-01-30
> This package is missing snoopprecompile: that seems to be a major oversight given its mostly a compile time burdened library?

I don't think so, because this package seems to always compile for every new thing it has to do.
It seems like it is severely over specializing.
But I did try adding a `@nospecialize`, which didn't seem to help much either, so perhaps I am wrong there/should try it more.

I'm speaking in complete ignorance about the internals of this library.
Just that I don't think that test script I ran should take anywhere close to 20 minutes. 5.5 minutes is better, but still way too long IMO.