@ChrisRackauckas 2023-01-30
> I don't think so, because this package seems to always compile for every new thing it has to do.
> It seems like it is severely over specializing.

Oh that's not good 😅 . I would assume that something based around printing like this would just use one set of unparameterized types over symbols, similar to writing a compiler. If that's not the case, I guess the good thing is that there is big room for improvement.