@chriselrod 2023-01-30
I think `max_methods=1` is generally safe and should be the defualt in Julia ([LoopVectorization.jl even uses it](https://github.com/JuliaSIMD/LoopVectorization.jl/blob/807675dc8ed40783891df1af790527f70102cfbb/src/LoopVectorization.jl#L3-L6) -- not because it makes a difference, but to emphasize that I think it is mostly harmless).

That's not the case for setting the opt-level to `1`, though.

So, I think these are all reasonable options with respect to this PR:
 - Merge as is
 - Drop the opt level change, merge only max_methods
 - Close