@korsbo 2023-01-31
Yeah, let's merge this. It might not do all the magic one would have hoped for but latency is painful while runtime is insignificant so anything that tries to balance the two more favourably is a good step.